// client/src/types/features-page.ts
export type FeaturesPage = {
  pageTitle: string;
  introText: string;

  feature1Title: string;
  feature1Description: string;
  feature1Category: 'Core' | 'Safety' | 'Social' | 'Profile';
  feature1IsNew: boolean;

  feature2Title: string;
  feature2Description: string;
  feature2Category: 'Core' | 'Safety' | 'Social' | 'Profile';
  feature2IsNew: boolean;

  feature3Title: string;
  feature3Description: string;
  feature3Category: 'Core' | 'Safety' | 'Social' | 'Profile';
  feature3IsNew: boolean;

  feature4Title: string;
  feature4Description: string;
  feature4Category: 'Core' | 'Safety' | 'Social' | 'Profile';
  feature4IsNew: boolean;

  feature5Title: string;
  feature5Description: string;
  feature5Category: 'Core' | 'Safety' | 'Social' | 'Profile';
  feature5IsNew: boolean;

  feature6Title: string;
  feature6Description: string;
  feature6Category: 'Core' | 'Safety' | 'Social' | 'Profile';
  feature6IsNew: boolean;
};